# Optimize kRelaxedSubstring LIKE pattern