… velox/benchmarks/basic/LikeBenchmark.cpp
… velox/docs/functions/presto/regexp.rst
… velox/functions/lib/Re2Functions.h
… velox/functions/lib/tests/Re2FunctionsTest.cpp