kRelaxedSubstring LIKE pattern is patterns that contains '%' but it only occurs at the beginning and end of the pattern, and it also contains '_'. e.g. '%Hello_Velox%'.

The algorithm is as follows:

kRelaxedSubstring pattern is consists of two kinds of sub patterns: `kLiteralString`, `kSingleCharWildcard`, we will use notation `L0`, `Li`, `S0`, `Si` etc to describe them in the following, e.g. `<L0><S0>` is a pattern like: `%hello__%`, while `<S0><L0>` is a pattern like `%__hello%`

1. We start by set the target sub-pattern to search as `L0`(currentSubPattern = L0).
2. We search for `currentSubPattern` in the input and record the result in `currentMatchingOffsetInInput`.
3. If it is not found(`currentMatchingOffsetInInput < 0`), algorithm finish with return result `false`.
4. If it is found, we then match the right part sub-patterns
  a. If it fails at some sub-pattern `Li`, we update `currentSubPattern` as `Li`, update `currentMatchingOffsetInInput` as the next byte where we failed the matching, and jump to step 2.
  b. If it fails at some sub-pattern `Si`, algorithm finish with return result `false`.
5. If the right sub-patterns matches, we continue match left sub-patterns.
  a. If it matches, algorithm finish with return result `true`.
  b. If it failed at some sub-pattern `Lj`, we update `currentSubPattern` as `Lj`, update `currentMatchingOffsetInInput` as the next byte where we failed the matching, and jump to step 2.
  c. If it failed at some sub-pattern `Sj`, we keep `currentSubPattern` not changed, and increment `currentMatchingOffsetInInput` by one byte, and jump to step 2.

 We designed benchmarks which consists of 4 tests:

- easy: no noise in the input.
- normal: many noise in the input, but easy to have fast path.
- hard_ascii: many noise in the pure ASCII input.
- hard_unicode: many noise in the Unicode input.

Before this optimization:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
relaxed_substring##easy                                      1.06s   945.37m
relaxed_substring##normal                                 994.73ms      1.01
relaxed_substring##hard_ascii                                4.67s   213.91m
relaxed_substring##hard_unicode                              7.08s   141.33m
```

After this optimization:

```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
relaxed_substring##easy                                     5.73ms    174.64
relaxed_substring##normal                                   6.93ms    144.36
relaxed_substring##hard_ascii                                1.20s   836.17m
relaxed_substring##hard_unicode                              1.79s   559.71m
```

So in summary: for normal input this optimization gives us 180x speedup, for hard pure ASCII or Unicode input it gives us 4x speedup.