@xumingming 2024-02-02
@mbasmanova The code has some conflicts I have to resolve first. The algorithm is as follows:

1. Start by searching the first literal sub-pattern in the input.
2. If we find the literal sub-pattern, we match the left & right part of the whole pattern.
3. If we failed matching any sub-pattern, we continue by searching the failed literal sub-pattern.

The key here is the 3rd step, we use failed sub-pattern to help us skip a lot of characters in the input.