@xumingming 2024-02-03
@mbasmanova Take the following test as an example:

```
  testLike(
      "a_ba_ba_ba_ba_ba_ba_ba_ba_ba_ba_ba_ba_ba_ba_ba_ba_b_cdef",
      "%a_b_c%",
      true);
```

The sub-patterns we are trying to match is:

- a
- _
- b
- _
- c

1. The first literal sub-pattern is `a`, so we start by searching for `a` in the input.
2. We find it at index: 0
3. We continue matching the right part of the whole pattern(_b_c), and we will fail at the sub-pattern `c`
4. Now we search for `c` in the input, we find it at the end of input.
5. Now we try to match the whole pattern at the index where we find c.
6. `c` is the right-est pattern in the whole pattern, so we no need to match the `the right part of the pattern`.
7. We match the left part of the pattern(a_b_), yeah, it matches.

We can see that by searching the failed sub-pattern `c` helped us skip a lot of characters. 

But there are cases which has a lot of noise which can not skip a lot of characters like:

```
  // There are a lot 'noise': a_b__c which is similar to our pattern: a_b_c, it
  // is hard for kRelaxedSubstring.
  testLike(
      "a_b__ca_b__ca_b__ca_b__ca_b__ca_b__ca_b__ca_b__ca_b_cdef",
      "%a_b_c%",
      true);
```