@mbasmanova 2024-03-01
@xumingming James, would you update documentation to include this new pattern?

https://facebookincubator.github.io/velox/functions/presto/regexp.html