@mbasmanova 2024-03-01
@xumingming James,

the explanation of the algorithm in https://github.com/facebookincubator/velox/pull/8581?fbclid=IwAR31kOxadJ2WzYCpDdCzmM2UzQVGOFUQAPKQwGK4GTBiFUFYalSoro73NNY#issuecomment-1925021912 is very nice and helpful. Would you include some version of it in the PR description as well?