@xumingming 2024-03-05
@mbasmanova I made the following changes(all your above questions should be answered except the ones for docs):

1. Replaced std::pair<bool, int32_t> with `RelaxedPatternMatchResult` which contains `success`, `failed`, `failedForTooShortInput`, `match`, `tooShortInputToMatch` etc to make the code more easier to read.
2. Removed the duplicated comments which are now added in `RelaxedPatternMatchResult`.
3. Introduced a simple method: `isRightMostSubPattern` to make the code more readable. You suggested to change the control flow of `matchRelaxedSubstringPattern`, but I find that after all these helper method introduced, the code is simple enough, if you insist, I can try change the control flow(I think it might make the code more complex).
4. For the `right part partial match, left match failed` case, you are right, we need to add backtrack, otherwise there will be bugs, I added one case and detected it:

```c++
  // This one needs backtrack(in our implementation).
  testLike("a_b_aab_b", "%a__b%", true);
```

I added backtrack merchanism for this kind of case:

```
    // matchResult.cursor() is where we failed matching the sub-pattern, we
    // start from `matchResult.cursor() + 1` to search for the sub-pattern.
    currentMatchingOffsetInInput = matchResult.cursor() + 1;
```

Now `matchResult.cursor()` returns the cursor where we failed matching, so when matching left part of the pattern, this will give us a backtrack.

5. For your question: `Is this the only place where we advance currentMatchingOffsetInInput? How do we make sure we always advance and never get stuck?` Althrough we added backtrace which might cause the `currentMatchingOffsetInInput` going backwards, but remember that `currentMatchingOffsetInInput` is a cursor for the failed matching sub-pattern, if we looking from the first sub-pattern point of view, the cusor is always advancing, so we will not get stuck, because when search the failed sub-pattern, we are always search forward.