@xumingming 2024-03-05
> @xumingming James, would you update documentation to include this new pattern?
> 
> https://facebookincubator.github.io/velox/functions/presto/regexp.html

This doc currently seems has not mentioned the patterns that we have fast path. Are you suggesting that we add one paragraph to describe the patterns we have fast path?