@mbasmanova 2024-03-05
> This doc currently seems has not mentioned the patterns that we have fast path. 

I see the following: "Patterns ‘aaa’, ‘aaa%’, ‘%aaa’, where ‘aaa’ contains only regular characters and ‘_’ wildcards are evaluated without using regular expressions." We can update that sentence to include the pattern supported in this PR.