@mbasmanova 2024-03-05
> Added algorithm description in the PR description.

I see an example that shows how the algorithm works, but I don't see a formal proof that this algorithm works correctly in all cases. The description also doesn't clarify from which position we are starting next search and why.